backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
pydantic>=2.6
django-environ>=0.11
requests>=2.31
polyline>=2.0
pytest>=8.0
pytest-django>=4.8
